refactor: consolidate EventSidebar and standardize SectionHeader props
- Refactored `EventCard` to support linking to event details via a new `slug` prop.
- Consolidated duplicate `EventSidebar` implementations into a single shared component in `src/components/ui/`.
- Renamed `eyebrow` prop to `label` in `SectionHeader` for consistency with `PageHeader`.
- Refactored `EventHero` for better design system compliance using layout primitives.
- Updated `useHome` hook to provide event slugs for home page cards.
- Verified changes with Vitest, Playwright E2E tests, and manual visual inspection.
- Updated visual snapshots to reflect functional link additions on the dashboard.
