+++ src/components/ui/SectionHeader.tsx
@@ -2,7 +2,7 @@ import { Link } from 'react-router-dom';
 import { Stack, Text } from '@/layouts/Primitives';
 
 interface SectionHeaderProps {
-  eyebrow: string;
+  label: string;
   title: string;
   link?: {
     text: string;
@@ -10,12 +10,12 @@ interface SectionHeaderProps {
   };
 }
 
-export function SectionHeader({ eyebrow, title, link }: SectionHeaderProps) {
+export function SectionHeader({ label, title, link }: SectionHeaderProps) {
   return (
     <Stack direction="row" align="end" justify="between" marginBottom={4}>
       <Stack direction="col" gap={1}>
         <Text variant="mono" size="xs" color="dim" uppercase tracking="widest">
-          {eyebrow}
+          {label}
         </Text>
         <Text as="h3" size="3xl" weight="font-black" className="text-accent-navy">
           {title}
