+++ src/features/dashboard/useHome.ts
@@ -50,6 +50,7 @@ export function useHome() {
       name: event.title,
       location: event.location,
       schedule: event.schedule,
+      slug: event.slug,
     })),
     dancerPaths,
     hirePaths,
